move files to serverless folder
